GDPR, odber noviniek, hash hesla

+++ delete.php
@@ -1,6 +1,16 @@
 <?php
 include 'config.php'; 
 $id = $_GET['id'];
-$query = "DELETE FROM produkt WHERE idProduct = $id";
-mysqli_query($conn, $query);
-header('location: products.php');
+$type = $_GET['type'];
+
+if ($type = 'recenzia') {
+    $query = "DELETE FROM recenzie WHERE idRecenzia = $id";
+    mysqli_query($conn, $query);
+    header('location: recenzie.php');
+}
+
+if ($type = 'produkt') {
+    $query = "DELETE FROM produkt WHERE idProduct = $id";
+    mysqli_query($conn, $query);
+    header('location: products.php');
+}
